+++ src/core/best-practices.js
@@ -3,8 +3,8 @@
 // Handles the marking up of best practices, and can generate a summary of all of them.
 // The summary is generated if there is a section in the document with ID bp-summary.
 // Best practices are marked up with span.practicelab.
-import { addId, makeSafeCopy } from "./utils.js";
-import { lang as defaultLang, getIntlData } from "../core/l10n.js";
+import { addId, getIntlData, makeSafeCopy } from "./utils.js";
+import { lang as defaultLang } from "../core/l10n.js";
 import { hyperHTML } from "./import-maps.js";
 import { pub } from "./pubsubhub.js";
