+++ src/core/figures.js
@@ -4,8 +4,13 @@
 // Adds width and height to images, if they are missing.
 // Generates a Table of Figures wherever there is a #tof element.
 
-import { addId, renameElement, showInlineWarning, wrapInner } from "./utils.js";
-import { getIntlData } from "../core/l10n.js";
+import {
+  addId,
+  getIntlData,
+  renameElement,
+  showInlineWarning,
+  wrapInner,
+} from "./utils.js";
 import { hyperHTML } from "./import-maps.js";
 
 export const name = "core/figures";
