+++ src/core/render-biblio.js
@@ -2,9 +2,8 @@
 // Module core/render-biblio
 // renders the biblio data pre-processed in core/biblio
 
-import { addId } from "./utils.js";
+import { addId, getIntlData } from "./utils.js";
 import { biblio } from "./biblio.js";
-import { getIntlData } from "../core/l10n.js";
 import { hyperHTML } from "./import-maps.js";
 import { pub } from "./pubsubhub.js";
